Update AppContext.tsx
Fixed issue related to authentication.  getProfilePhoto on line 106 was returning 404 in cases where users don't have an avatar, which cause caught by the catch statement, meaning the remainder of the code in the try statement after line 106 was never executed. Also setAuthenticated was never being called in that same try catch stement, so I added that. Also accessToken and sessionId were never set on initial login, so I set added that, and wrapped in a try catch statement just incase.